Add stepSize rounding, periodic sync, Telegram daily report, and margin update refactor
- Load and log exchangeInfo specs (stepSize/tickSize/minQty) at startup
- Round DCA fill amounts with stepSize to prevent SL/TP quantity mismatch
- Add round_amount() safety guard in _place_sl_order/_place_tp_order
- Add periodic_sync() to detect TP/SL fills missed by WebSocket
- Cancel orphaned algo orders during initialization
- Defer margin+params update to 5-min cycle via _needs_update flag
- Fix position cost calculation (use notional instead of margin)
- Add Telegram daily report (main_daily_report.py) with PM2 cron
- Update ecosystem.config.js: replace dca-monitor with dca-report

+++ src/common/data_manager.py
@@ -176,7 +176,7 @@ class DataManager:
     def fetch_data(
         symbol: str,
         timeframe: str = "15m",
-        years: int = 4,
+        years: float = 4,
         start_date_str: Optional[str] = None,
         end_date_str: Optional[str] = None,
         force_refresh: bool = False,
